+++ spec/ig_markets/cli/main_spec.rb
@@ -15,12 +15,12 @@ describe IGMarkets::CLI::Main do
     IGMarkets::CLI::Main.begin_session(cli.options) { |dealing_platform| }
   end
 
-  it 'reports a standard error' do
-    expect(dealing_platform).to receive(:sign_in).and_raise('test')
+  it 'reports an argument error' do
+    expect(dealing_platform).to receive(:sign_in).and_raise(ArgumentError, 'test')
 
     expect do
       IGMarkets::CLI::Main.begin_session(cli.options) { |dealing_platform| }
-    end.to output("Error: test\n").to_stderr.and raise_error(SystemExit)
+    end.to output("Argument error: test\n").to_stderr.and raise_error(SystemExit)
   end
 
   it 'reports a request failure' do
@@ -28,7 +28,7 @@ describe IGMarkets::CLI::Main do
 
     expect do
       IGMarkets::CLI::Main.begin_session(cli.options) { |dealing_platform| }
-    end.to output("Request failed: test\n").to_stderr.and raise_error(SystemExit)
+    end.to output("Request error: test\n").to_stderr.and raise_error(SystemExit)
   end
 
   it 'reports a deal confirmation' do